move section containers to utils
