+++ frontend/src/components/utils/SectionContainer.vue
