+++ frontend/src/components/utils/SectionContainerItem.vue
