+++ frontend/src/pages/trainer/TrainerPage.vue
@@ -6,8 +6,8 @@ import NameLink from '@/components/utils/NameLink.vue';
 import MainButton from '@/components/utils/MainButton.vue';
 import { ref } from 'vue';
 import { Admin, CourseInfo, SessionInfo } from '@gym-manager/models';
-import SectionContainer from '@/components/SectionContainer.vue';
-import SectionContainerItem from '@/components/SectionContainerItem.vue';
+import SectionContainer from '@/components/utils/SectionContainer.vue';
+import SectionContainerItem from '@/components/utils/SectionContainerItem.vue';
 
 const store = useUserStore();
