+++ frontend/src/pages/user/UserPage.vue
@@ -7,8 +7,8 @@ import MainButton from '@/components/utils/MainButton.vue';
 import { CourseInfo, SessionInfo, Trainer } from '@gym-manager/models';
 import { ref } from 'vue';
 import { useModalsStore } from '@/store/modals';
-import SectionContainer from '@/components/SectionContainer.vue';
-import SectionContainerItem from '@/components/SectionContainerItem.vue';
+import SectionContainer from '@/components/utils/SectionContainer.vue';
+import SectionContainerItem from '@/components/utils/SectionContainerItem.vue';
 
 
 const store = useUserStore();
